refactor: relocate routes
