+++ src/main/java/routes/Route.java
@@ -1,4 +1,4 @@
-package controllers;
+package routes;
 
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
